@@ -13,21 +13,21 @@
     "e2e": "ng e2e",
     "dev:ssr": "ng run app:serve-ssr",
     "serve:ssr": "node dist/app/server/main.js",
-    "build:ssr": "ng build --configuration=production && ng run app:server:production",
+    "build:ssr": "ng build && ng run app:server",
     "prerender": "ng run app:prerender",
     "clean:build:android": "node -e \"const fs = require('fs'); const path = require('path'); const deleteFiles = (directory, extensions) => { fs.readdir(directory, (err, files) => { if (err) throw err; for (const file of files) { const filePath = path.join(directory, file); fs.stat(filePath, (err, stat) => { if (err) throw err; if (stat.isDirectory()) { deleteFiles(filePath, extensions); } else if (extensions.some(ext => filePath.endsWith(ext))) { fs.unlink(filePath, err => { if (err) throw err; }); } }); } }); }; const distPath = path.join(__dirname, 'android/app/src/main/assets/public'); deleteFiles(distPath, ['.gz', '.br', '.map']);\""
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "^17.3.11",
-    "@angular/core": "^17.3.11",
-    "@angular/forms": "^17.3.11",
-    "@angular/platform-browser": "^17.3.11",
-    "@angular/platform-browser-dynamic": "^17.3.11",
-    "@angular/platform-server": "^17.3.11",
-    "@angular/router": "^17.3.11",
-    "@angular/service-worker": "^17.3.11",
-    "@angular/ssr": "^17.3.8",
+    "@angular/common": "^18.1.0",
+    "@angular/core": "^18.1.0",
+    "@angular/forms": "^18.1.0",
+    "@angular/platform-browser": "^18.1.0",
+    "@angular/platform-browser-dynamic": "^18.1.0",
+    "@angular/platform-server": "^18.1.0",
+    "@angular/router": "^18.1.0",
+    "@angular/service-worker": "^18.1.0",
+    "@angular/ssr": "^18.1.0",
     "@capacitor-firebase/authentication": "^6.0.0",
     "@capacitor/android": "6.1.0",
     "@capacitor/app": "6.0.0",
@@ -45,14 +45,12 @@
     "@ionic/core": "^7.0.3",
     "@ionic/storage-angular": "^4.0.0",
     "@ngneat/transloco": "^4.0.0",
-    "@ngrx/component-store": "^17.2.0",
-    "@ngrx/effects": "^17.2.0",
-    "@ngrx/entity": "^17.2.0",
-    "@ngrx/router-store": "^17.2.0",
-    "@ngrx/store": "^17.2.0",
-    "@ngrx/store-devtools": "^17.2.0",
-    "@sentry/angular-ivy": "^7.54.0",
-    "@sentry/capacitor": "^0.18.0",
+    "@ngrx/component-store": "^18.0.1",
+    "@ngrx/effects": "^18.0.1",
+    "@ngrx/entity": "^18.0.1",
+    "@ngrx/router-store": "^18.0.1",
+    "@ngrx/store": "^18.0.1",
+    "@ngrx/store-devtools": "^18.0.1",
     "@types/lodash": "^4.14.186",
     "animate.css": "^4.1.1",
     "chart.js": "^3.8.0",
@@ -61,7 +59,7 @@
     "chess.js": "^0.12.1",
     "cm-chessboard": "^7.7.1",
     "date-fns": "^2.28.0",
-    "express": "^4.15.2",
+    "express": "^4.18.2",
     "firebase": "^10.12.0",
     "howler": "^2.2.3",
     "lodash": "^4.17.21",
@@ -70,20 +68,20 @@
     "zone.js": "~0.14.7"
   },
   "devDependencies": {
-    "@angular-builders/custom-webpack": "^17.0.2",
-    "@angular-devkit/build-angular": "^17.3.8",
+    "@angular-builders/custom-webpack": "^18.0.0",
+    "@angular-devkit/build-angular": "^18.1.0",
     "@angular-eslint/builder": "~15.1.0",
     "@angular-eslint/eslint-plugin": "~15.1.0",
     "@angular-eslint/eslint-plugin-template": "~15.1.0",
     "@angular-eslint/template-parser": "~15.1.0",
-    "@angular/cli": "^17.3.8",
-    "@angular/compiler": "^17.3.11",
-    "@angular/compiler-cli": "^17.3.11",
-    "@angular/language-service": "^17.3.11",
+    "@angular/cli": "^18.1.0",
+    "@angular/compiler": "^18.1.0",
+    "@angular/compiler-cli": "^18.1.0",
+    "@angular/language-service": "^18.1.0",
     "@capacitor/cli": "^6.1.0",
     "@ionic/angular-toolkit": "^5.0.0",
     "@ngrx/schematics": "^17.2.0",
-    "@types/express": "^4.17.0",
+    "@types/express": "^4.17.17",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^12.11.1",
