@@ -2,7 +2,7 @@ import { APP_INITIALIZER, ErrorHandler, NgModule, isDevMode } from '@angular/cor
 import { BrowserModule } from '@angular/platform-browser';
 import { Router, RouteReuseStrategy } from '@angular/router';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { HttpClientModule } from '@angular/common/http';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 
 import { IonicModule, IonicRouteStrategy } from '@ionic/angular';
 import { IonicStorageModule } from '@ionic/storage-angular';
@@ -27,25 +27,7 @@ import { CustomRouterStateSerializer } from '@redux/states/router.state';
 import * as fromEffects from '@redux/effects';
 import * as fromGuards from '@guards/index';
 
-import * as Sentry from '@sentry/capacitor';
-import * as SentryAngular from '@sentry/angular-ivy';
 
-Sentry.init(
-  {
-    dsn: 'https://4716d5ae6b19528674465cfafd357307@o4507017740681216.ingest.us.sentry.io/4507019628314624',
-    // To set your release and dist versions
-    release: 'chesscolate@' + environment.version,
-    dist: '1',
-    // Set tracesSampleRate to 1.0 to capture 100% of transactions for performance monitoring.
-    // We recommend adjusting this value in production.
-    tracesSampleRate: environment.production ? 0.25 : 1,
-    environment: environment.production ? 'production' : 'development',
-    ignoreErrors: [
-    ],
-  },
-  // Forward the init method from @sentry/angular
-  SentryAngular.init
-);
 
 
 import { TranslocoRootModule } from './transloco-root.module';
@@ -71,8 +53,7 @@ if (environment.production) {
 
 @NgModule({
   declarations: [AppComponent],
-  imports: [
-    BrowserModule,
+  bootstrap: [AppComponent], imports: [BrowserModule,
     IonicModule.forRoot(),
     IonicStorageModule.forRoot(),
     AppRoutingModule,
@@ -83,8 +64,6 @@ if (environment.production) {
     StoreModule.forRoot(appReducers),
     ...devImports,
     EffectsModule.forRoot(fromEffects.EFFECTS),
-    // SocketIoModule.forRoot(config),
-    HttpClientModule,
     TranslocoRootModule,
     FormsModule,
     ReactiveFormsModule,
@@ -94,27 +73,10 @@ if (environment.production) {
       // Register the ServiceWorker as soon as the application is stable
       // or after 30 seconds (whichever comes first).
       registrationStrategy: 'registerWhenStable:30000'
-    })
-  ],
-  providers: [
-    PROVIDERS,
-    { provide: RouteReuseStrategy, useClass: IonicRouteStrategy },
-    {
-      provide: ErrorHandler,
-      // Attach the Sentry ErrorHandler
-      useValue: SentryAngular.createErrorHandler(),
-    },
-    {
-      provide: SentryAngular.TraceService,
-      deps: [Router],
-    },
-    {
-      provide: APP_INITIALIZER,
-      useFactory: () => () => { },
-      deps: [SentryAngular.TraceService],
-      multi: true,
-    },
-  ],
-  bootstrap: [AppComponent]
+    })], providers: [
+      PROVIDERS,
+      { provide: RouteReuseStrategy, useClass: IonicRouteStrategy },
+      provideHttpClient(withInterceptorsFromDi()),
+    ]
 })
 export class AppModule { }
